move generic content components out of views
