+++ src/web/components/Content/style.scss
@@ -1,0 +1,21 @@
+.modal {
+  :global(.btn-info) {
+    background-color: #46cba0;
+    border-color: #46cba0;
+
+    &:hover {
+      background-color: #40bd94;
+      border-color: #40bd94;
+    }
+  }
+}
+
+.cancel {
+  position: absolute;
+  left: 90px;
+  bottom: 24px;
+  height: 34px;
+  padding-right: 12px;
+  padding-left: 12px;
+  min-width: 70px;
+}
