+++ src/web/components/Layout/index.jsx
@@ -12,7 +12,7 @@ import HelpButton from './HelpButton'
 import LicenseComponent from '~/components/License'
 import AboutComponent from '~/components/About'
 import GuidedTour from '~/components/Tour'
-import SelectContent from '~/views/Content/Select'
+import SelectContent from '~/components/Content/Select'
 
 import PluginInjectionSite from '~/components/PluginInjectionSite'
